requirements.txt
flask
discord.py
requests
beautifulsoup4